+++ apps-client/ee/free/frames/signup.js
@@ -470,6 +470,7 @@ app.routes.frame('/cadastrar', function (params, data) {
                         }
                     } else {
                         var token = response.token;
+                        app.event('cadastrar');
                         app.close({token : token});
                     }
                 });